requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-CORS==4.0.0
Werkzeug==2.3.7
PyPDF2==3.0.1
Pillow==10.0.1
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.8.0
gunicorn==21.2.0
